+++ js/main.js
@@ -1,50 +1,5 @@
-const COMMENTS_NAMES = ['James', 'Julia', 'Alex', 'Michel', 'Grag', 'John'];
-const MESSAGES = ['Всё отлично!', 'В целом всё неплохо. Но не всё.', 'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.', 'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают.', 'Как можно было поймать такой неудачный момент?!'];
-const DESCRIPTIONS = ['Хороший ракурс', 'Отметили так отметили', 'No comments', 'Закат на Балли', 'Утро в Сочи', 'Милая фотка', 'Огонь'];
-const MAX_POST = 25;
-const AVATAR_SRC = 'img/avatar-';
-const AVATAR_FORMAT = '.svg';
-const PHOTO_SRC = 'photos/';
-const PHOTO_FORMAT = '.jpg';
-const MIN_LIKE = 15;
-const MAX_LIKE = 200;
+import {MAX_POST} from './render-moks-data';
+import {getArrayMocks} from './render-moks-data';
 
-
-// Получает положительное рандомное целое число > или = 0
-const getRandomInt = (min = 0, max = 0) => ((min >= 0) && (min < max)) ? Math.floor(Math.random() * (max - min + 1)) + min : 0;
-
-//  Проверяет длину коментария
-const checkCommentLength = (str = '', maxCommentLength = 140) => str.length <= maxCommentLength;
-
-const getUnquiArray = (count = 0) => {
-  const result = new Set();
-  while (result.size < count) {
-    result.add(getRandomInt(1,300));
-  }
-  return Array.from(result);
-};
-
-const getArrayMocks = (count = 0) => {
-  const result = [];
-  const moksCommentsArray = [];
-  const unquiElements = getUnquiArray(MAX_POST);
-  for (let index = 0; index < count; index++) {
-    moksCommentsArray.push({
-      id: unquiElements[index],
-      avatar: AVATAR_SRC + getRandomInt(1, 6) + AVATAR_FORMAT,
-      message: checkCommentLength(MESSAGES[getRandomInt(0, MESSAGES.length - 1)]),
-      name: COMMENTS_NAMES[getRandomInt(0, COMMENTS_NAMES.length - 1)],
-    });
-    result.push({
-      id: index + 1,
-      url: `${PHOTO_SRC}${index + 1}${PHOTO_FORMAT}`,
-      description: DESCRIPTIONS[getRandomInt(0, DESCRIPTIONS.length - 1)],
-      likes: getRandomInt(MIN_LIKE, MAX_LIKE),
-      comments: moksCommentsArray[index],
-    });
-  } return result;
-};
 const moksPostsArray = getArrayMocks(MAX_POST);
 moksPostsArray.length;
